Part of WOR-369: split tests/test_watcher.py into themed siblings (dispatch_loop, signal_handling, worker_lifecycle) (#778)
